ref: moved Sidebar elements into /elements directory
