+++ src/components/root/Navigation/desktop/DesktopSideBar.tsx
@@ -1,9 +1,9 @@
 import { twMerge as tw } from 'tailwind-merge'
-import RenderSideBarItems from '@/components/root/Navigation/RenderSideBarItems'
 import SidebarHoverabilityDetection from '@/components/root/Navigation/desktop/SidebarHoverabilityDetection'
 import ThemeSwitcher from '@/components/root/ThemeSwitcher'
 import DesktopSidebarDialog from '@/components/root/Navigation/desktop/DesktopSidebarDialog'
-import SidebarUserBanner from '@/components/root/Navigation/user/SidebarUserBanner'
+import SidebarUserBanner from '@/components/root/Navigation/elements/SidebarUserBanner'
+import RenderSideBarItems from '@/components/root/Navigation/elements/RenderSideBarItems'
 
 export const DesktopSidebar = ({ className, children }: { children: React.ReactNode; className?: string }) => {
   return (
